refactor(haptics): address code review — de-dup helpers, test toggle wiring
- Extract shared SettingsToggleRow into id.homebase.core.widget, replacing the
  private copies in NotificationSettingsScreen and AppearanceSettingsScreen. A
  contentPadding param defaults to the standard row inset so Notification's call
  sites are unchanged; Appearance passes PaddingValues(0.dp).
- Promote InMemorySettings to a shared commonTest fixture, removing the duplicate
  copies in ReactionMenuTest and UserPreferencesScrollAnchorTest.
- Add AppearanceSettingsUiTest.hapticFeedbackToggleFiresAction covering the
  Switch -> HapticsEnabledChanged wiring (previously untested).
