+++ homebase-core/src/commonMain/kotlin/id/homebase/core/ui/screens/appearance/AppearanceSettingsScreen.kt
@@ -2,10 +2,9 @@ package id.homebase.core.ui.screens.appearance
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.consumeWindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
@@ -17,15 +16,12 @@ import androidx.compose.material.icons.filled.LightMode
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.testTag
@@ -36,6 +32,7 @@ import id.homebase.core.settings.ThemeState
 import id.homebase.core.settings.setPlatformSystemLocale
 import id.homebase.core.widget.SettingsClickableRow
 import id.homebase.core.widget.SettingsRowItemData
+import id.homebase.core.widget.SettingsToggleRow
 import id.homebase.resources.MR
 import id.homebase.resources.language
 import id.homebase.resources.language_danish
@@ -150,29 +147,9 @@ fun AppearanceSettingsUi(
                 label = stringResource(MR.string.settings_haptic_feedback),
                 checked = uiState.hapticsEnabled,
                 onCheckedChange = { onAction(AppearanceSettingsUiAction.HapticsEnabledChanged(it)) },
+                contentPadding = PaddingValues(0.dp),
             )
         }
-    }
-}
-
-@Composable
-private fun SettingsToggleRow(
-    label: String,
-    checked: Boolean,
-    onCheckedChange: (Boolean) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Row(
-        modifier = modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Text(
-            text = label,
-            style = MaterialTheme.typography.bodyLarge,
-            modifier = Modifier.weight(1f),
-        )
-        Switch(checked = checked, onCheckedChange = onCheckedChange)
     }
 }
